splitting dao file
