+++ code/server/routes/controller/student.js
@@ -1,4 +1,6 @@
-const dao = require('../../dao');
+const daoStudent = require('../../daoStudent');
+const daoTeacher = require('../../daoTeacher');
+const daoGeneral = require('../../daoGeneral');
 const dayjs = require('dayjs');
 
 const insertNewApplication = async (req, res) => {
@@ -8,21 +10,21 @@ const insertNewApplication = async (req, res) => {
     const timestamp = dayjs().format("DD/MM/YYYY HH:mm:ss");
     const status = 'pending';
 
-    if(req.user.id !== studentId){
-        return res.status(422).json({error:"the student who is sending the application is not the logged in one"});
+    if (req.user.id !== studentId) {
+        return res.status(422).json({ error: "the student who is sending the application is not the logged in one" });
     }
     try {
-        const changes = await dao.addApplicationForThesis(proposalId, studentId, timestamp, status,teacherId);
+        const changes = await daoStudent.addApplicationForThesis(proposalId, studentId, timestamp, status, teacherId);
         return res.status(201).json(changes);
-    } catch(e) {
-        if(e.message.includes("SQLITE_CONSTRAINT")) {
-            e.message = "Application already submitted, wait for professor response";       
+    } catch (e) {
+        if (e.message.includes("SQLITE_CONSTRAINT")) {
+            e.message = "Application already submitted, wait for professor response";
         }
         return res.status(500).json(e.message);
     }
 }
 
-const getApplicationsForStudent = async (req,res) => {
+const getApplicationsForStudent = async (req, res) => {
     // since I dont know how to log in in the testapi.http file using the new SAML strategy,
     // for the moment I'm passing as param the studentId (should be found in the session cookie)
     //const studentId = req.params.studentId
@@ -30,23 +32,23 @@ const getApplicationsForStudent = async (req,res) => {
     // ideally, the API should be protected, and the studentId should be retrieved from the user object
     const studentId = req.user.id
     try {
-        const applications = await dao.getApplicationsForStudent(studentId);
-        if(applications.error) {
+        const applications = await daoStudent.getApplicationsForStudent(studentId);
+        if (applications.error) {
             return res.status(404).json(applications);
         } else {
             const enhancedApplications = [];
             //add the 2 fields with details to the object
             for (const appl of applications) {
-                const teacherInfo = await dao.getTeacherById(appl.teacherId);
-                const thesisInfo = await dao.getThesisProposalById(appl.thesisId, appl.status)
+                const teacherInfo = await daoTeacher.getTeacherById(appl.teacherId);
+                const thesisInfo = await daoGeneral.getThesisProposalById(appl.thesisId, appl.status)
                     .then(t => {
-                        if(t.error){
-                            return dao.getThesisProposalById(appl.thesisId)
+                        if (t.error) {
+                            return daoGeneral.getThesisProposalById(appl.thesisId)
                         } else {
                             return t;
                         }
                     });
-                const studentInfo = await dao.getStudentById(appl.studentId);
+                const studentInfo = await daoStudent.getStudentById(appl.studentId);
 
                 enhancedApplications.push({
                     ...appl,
@@ -59,19 +61,19 @@ const getApplicationsForStudent = async (req,res) => {
                 enhancedApplications
             });
         }
-    } catch(e) {
+    } catch (e) {
         return res.status(500).json(e.message);
     }
 }
 
-const getThesisProposals = async (req,res) => {
+const getThesisProposals = async (req, res) => {
     //same principle for the getApplications: for manual testing purposes at the moment the degree
     //code is taken as param; ideally, it should be taken from the req.user object
     //const studentCourse = req.params.degreeCode
     const studentCourse = req.user.degree_code
     try {
-        const proposals = await dao.getThesisProposals(studentCourse);
-        if(proposals.error){
+        const proposals = await daoStudent.getThesisProposalsByDegree(studentCourse);
+        if (proposals.error) {
             return res.status(404).json(proposals);
         } else {
             return res.status(200).json(proposals);
@@ -80,8 +82,8 @@ const getThesisProposals = async (req,res) => {
         return res.status(500).json(e.message);
     }
 }
- module.exports = {
-     insertNewApplication,
-     getApplicationsForStudent,
-     getThesisProposals
- };
+module.exports = {
+    insertNewApplication,
+    getApplicationsForStudent,
+    getThesisProposals
+};
