+++ code/server/routes/utils/utils.js
@@ -1,7 +1,11 @@
-const dao = require('../../dao');
-const dayjs = require('dayjs');;
-const customParseFormat = require("dayjs/plugin/customParseFormat")
+const daoStudent = require('../../daoStudent');
+const daoTeacher = require('../../daoTeacher');
+const daoGeneral = require('../../daoGeneral');
+const daoUtils= require('../../daoUtils');
 const models = require('../../model');
+const dayjs = require('dayjs');
+const customParseFormat = require("dayjs/plugin/customParseFormat")
+
 
 dayjs.extend(customParseFormat);
 
@@ -9,9 +13,9 @@ const rearrangeProposals = async (req,res) => {
     const selectedTimestamp = req.body.selectedTimestamp;
     try {
         let counterMovedProposals = 0;
-        let expiredProposals = await dao.getExpiredProposals(selectedTimestamp);
-        let proposalsToRevive = await dao.getProposalsToRevive(selectedTimestamp);
-        const acceptedProposalsIds = await dao.getAcceptedProposalsIds()
+        let expiredProposals = await daoUtils.getExpiredProposals(selectedTimestamp);
+        let proposalsToRevive = await daoUtils.getProposalsToRevive(selectedTimestamp);
+        const acceptedProposalsIds = await daoUtils.getAcceptedProposalsIds()
             .then(ids => {
                 if(ids.error){
                     return [];
@@ -24,7 +28,7 @@ const rearrangeProposals = async (req,res) => {
             for (p of expiredProposals) { 
                 //archive expired proposals (insert into archived_thesis_proposals, delete from thesis proposals)
                 //also mark as expired all applications for expired proposals
-                await dao.archiveProposal(new models.ThesisProposal(
+                await daoTeacher.archiveProposal(new models.ThesisProposal(
                     p.id, //can be whatever, DB handles autoincrement id
                     p.title,
                     p.supervisor,
@@ -40,8 +44,8 @@ const rearrangeProposals = async (req,res) => {
                     p.cds.join(',')
                 ))
                     .then(() => {
-                        dao.deleteProposal(p.id);
-                        dao.updateApplicationsForExpiredProposals(p.id, p.supervisor.slice(0,6));
+                        daoTeacher.deleteProposal(p.id);
+                        daoUtils.updateApplicationsForExpiredProposals(p.id, p.supervisor.slice(0,6));
                         counterMovedProposals++;
                     })
                     .catch( (e) => {
@@ -59,7 +63,7 @@ const rearrangeProposals = async (req,res) => {
                 //NOTE: revive a proposal only if there isn't already an accepted application for that proposal, otherwise
                 //ignore it
                 if(!acceptedProposalsIds.includes(p.id)) {
-                    await dao.saveNewProposal(new models.ThesisProposal(
+                    await daoTeacher.saveNewProposal(new models.ThesisProposal(
                         p.id, //can be whatever, DB handles autoincrement id
                         p.title,
                         p.supervisor,
@@ -75,8 +79,8 @@ const rearrangeProposals = async (req,res) => {
                         p.cds.join(',')
                     ))
                         .then(() => {
-                            dao.reviveExpiredApplications(p.id)
-                                .then(dao.deleteProposalFromArchived(p.id))
+                            daoUtils.reviveExpiredApplications(p.id)
+                                .then(daoUtils.deleteProposalFromArchived(p.id))
                             counterMovedProposals++;
                         })
                         .catch ( (e) => {
@@ -87,7 +91,7 @@ const rearrangeProposals = async (req,res) => {
         }
 
         //clean up eventual applications that need to be put back as canceled
-        await dao.cancelApplicationsAfterClockChange();
+        await daoUtils.cancelApplicationsAfterClockChange();
         
         //if we get here, all went well
         return res.status(200).json(counterMovedProposals);
